% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mran.R
\name{cran_get_package_number_mran}
\alias{cran_get_package_number_mran}
\title{Get number of package on CRAN on a given date using MRAN}
\usage{
cran_get_package_number_mran(dates)
}
\arguments{
\item{dates}{A vector of dates. Either a character vector of the form "yyyy-mm-dd"
or a vector of class "Date".}
}
\value{
A data.frame with two volumns `date` and `n` the number of packages on CRAN
at that given `date`.
}
\description{
This function queries MRAN to retrieve the number of package on CRAN
on a given date.
}
\examples{
cran_get_package_number_mran(c("2018-04-10", "2020-03-19"))
}
