% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_cran_monthly_package_number.R
\docType{data}
\name{cran_monthly_package_number}
\alias{cran_monthly_package_number}
\title{cran_monthly_package_number}
\format{
A data frame with 297 rows and 2 variables:
\describe{
  \item{ date }{  Date }
  \item{ number_packages }{  numeric }
}
}
\source{
Source
}
\usage{
cran_monthly_package_number
}
\description{
Description.
}
\keyword{datasets}
