% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranology-package.R
\docType{package}
\name{cranology-package}
\alias{cranology}
\alias{cranology-package}
\title{cranology: The CRAN Chronology}
\description{
Scraping routines and datasets to monitor the evolution of the number of packages on CRAN.
}
\author{
\strong{Maintainer}: Antoine Languillaume \email{antoine@thinkr.fr} (\href{https://orcid.org/0000-0001-9843-5632}{ORCID})

Authors:
\itemize{
  \item Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
