% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranology.R
\name{cranology}
\alias{cranology}
\title{Get number of packages on CRAN at a given date}
\usage{
cranology(date)
}
\arguments{
\item{date}{A character string representing a date of the form yyyy-mm-dd.}
}
\value{
A data.frame with 2 columns \code{date} and \code{n} the number of packages on CRAN
at that date.
}
\description{
Query MRAN to Get number of packages on CRAN at a given date.
MRAN -> Microsoft snapshots of CRAN.
}
\examples{
cranology("2020-03-15")
}
