% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cran.R
\name{get_cran_monthly_package_number}
\alias{get_cran_monthly_package_number}
\title{Title}
\usage{
get_cran_monthly_package_number(
  cran_packages_history_df = cran_packages_history
)
}
\arguments{
\item{cran_packages_history_df}{A data.frame similar to cranology::cran_packages_history
which is the default value.}
}
\description{
Description
}
