% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mran.R
\name{get_package_number_cran}
\alias{get_package_number_cran}
\title{Title}
\usage{
get_package_number_cran(dates)
}
\arguments{
\item{dates}{A vector of dates.}
}
\value{
A data.frame
}
\description{
Description
}
