% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cran_monthly_package_number}
\alias{plot_cran_monthly_package_number}
\title{Plot monthly evolution of package number on CRAN}
\usage{
plot_cran_monthly_package_number()
}
\value{
A ggplot object
}
\description{
This function is a convenience tool to quickly draw a line
showing the evolution of packages number on CRAN since its
beginning. It uses the `cran_monthly_package_number` dataset.
}
\examples{
plot_cran_monthly_package_number()
}
