% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cran.R
\name{scrape_cran_history}
\alias{scrape_cran_history}
\title{Scrape CRAN packages history}
\usage{
scrape_cran_history()
}
\value{
A list of two data.frames: `cran_packages_history` and
`cran_monthly_package_number`.
}
\description{
This function is the workhorse of {cranology}. It scrapes
https://cran.rstudio.com and generates two datasets:
}
\details{
* `cran_packages_history`: A data.frame gathering information about every package
that has ever been on CRAN including the first release date the number
of versions released so far...
* `cran_monthly_package_number`: A data.frame holding the number of
packages available on CRAN since its beginning. Data is provided on
a montly basis.
}
\examples{
\dontrun{
scrape_cran_history()
}
}
