% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_cran_monthly_package_number.R
\docType{data}
\name{cran_monthly_package_number}
\alias{cran_monthly_package_number}
\title{cran_monthly_package_number}
\format{
A data frame with 297 rows and 2 variables:
\describe{
  \item{ date }{  [Date] Month of release.}
  \item{ number_packages }{  [numeric] Number of packages available on CRAN
  at that given date.}
}
}
\source{
https://cran.rstudio.com/src/contrib/ and https://mran.microsoft.com/
}
\usage{
cran_monthly_package_number
}
\description{
The evolution of the number of packages on CRAN since its beginning.
}
\keyword{datasets}
