% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_cran_packages_history.R
\docType{data}
\name{cran_packages_history}
\alias{cran_packages_history}
\title{cran_packages_history}
\format{
A data frame with 37459 rows and 10 variables:
\describe{
  \item{ file_name }{  character }
  \item{ date }{  POSIXct,POSIXt }
  \item{ time }{  character }
  \item{ size }{  character }
  \item{ package_name }{  character }
  \item{ last_archived }{  POSIXct,POSIXt }
  \item{ archive }{  logical }
  \item{ first_date }{  POSIXct,POSIXt }
  \item{ n_versions }{  integer }
  \item{ last_modified }{  POSIXct,POSIXt }
}
}
\source{
Source
}
\usage{
cran_packages_history
}
\description{
All packages ever available on CRAN as of 2022-04-01.
}
\keyword{datasets}
