% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mran.R
\name{mran}
\alias{mran}
\alias{get_package_number_mran}
\title{Get number of package on CRAN on a given date using MRAN}
\usage{
get_package_number_mran(dates, parallelize = FALSE)
}
\arguments{
\item{dates}{A vector of dates. Either a character vector of the form "yyyy-mm-dd"
or a vector of class "Date". All dates must be posterior to "2014-09-17", the
day of MRAN first CRAN snapshot.}

\item{parallelize}{A logical. If TRUE {furrr} is used to asynchronously scrap
mran.}
}
\value{
A data.frame with two columns `date` and `n` the number of packages on CRAN
at that given `date`.
}
\description{
This function queries MRAN to retrieve the number of package on CRAN
on a given date.
}
\examples{
get_package_number_mran(c("2018-04-10", "2020-03-19"))

}
