% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mran.R
\name{update_monthly_package_number}
\alias{update_monthly_package_number}
\title{Update `cran_monthly_package_number` dataset}
\usage{
update_monthly_package_number(
  cran_monthly_package_number_df,
  parallelize = FALSE
)
}
\arguments{
\item{cran_monthly_package_number_df}{A data.frame similar to
the `cran_monthly_package_number` dataset included within {cranology}.}

\item{parallelize}{A logical. If TRUE {furrr} is used to asynchronously scrap
mran.}
}
\description{
The creation `cran_monthly_package_number` using `scrape_cran()` is
a long process as the underlying scrapping operations are time consuming.
To more rapidly update `cran_monthly_package_number` it is easier to rely
on data from MRAN. This what this function does. It uses
#' `get_package_number_mran()` to quickly update the dataset.
}
\examples{
# Simulate `cran_monthly_package_number` update
date_lag <- 3
df <- cran_monthly_package_number[
  1:(nrow(cran_monthly_package_number) - date_lag),
]
update_monthly_package_number(
  cran_monthly_package_number_df = df
)

}
